fixed svelte and vue example
