+++ examples/svelte/src/main.js
@@ -1,11 +1,10 @@
-
 import App from './App.svelte';
 
-// var app
-// document.addEventListener("DOMContentLoaded", function(event) { 
-let app = new App({
-		target: document.body,	
-	})
-// })
+var app
+document.addEventListener("DOMContentLoaded", function() { 
+    app = new App({
+        target: document.body,	
+    })
+})
 
 export default app
