+++ examples/svelte_todo/package.json
@@ -8,7 +8,7 @@
     "url": "git+https://github.com/marcomq/nimview.git"
   },
   "scripts": {
-    "build": "rollup -c && mkdirp ./dist && inliner -m --videos --preserve-comments public/index.html > dist/index.html",
+    "build": "rollup -c  && inliner -m --videos --preserve-comments dist/index.html > dist/nimview.html",
     "devbuild": "rollup -c",
     "autobuild": "rollup -c -w",
     "dev": "run-p start:dev autobuild",
@@ -21,14 +21,13 @@
     "@babel/plugin-transform-runtime": "^7.14.3",
     "@babel/preset-env": "^7.14.4",
     "@babel/runtime": "^7.14.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-commonjs": "^17.0.0",
     "@tauri-apps/tauri-inliner": "^1.14.1",
-    "mkdirp": "^1.0.4",
     "npm-run-all": "^4.1.5",
     "rollup": "^2.50.4",
     "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-dev": "^1.1.3",
     "rollup-plugin-json": "^4.0.0",
