+++ examples/vue/babel.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   presets: [
-    '@vue/cli-plugin-babel/preset'
+    ['@vue/app', { useBuiltIns: 'usage' }]
   ]
 }
