+++ examples/vue/vue.config.js
@@ -5,5 +5,8 @@ module.exports = {
   devServer: {
     port: 5000, 
     proxy: 'http://127.0.0.1:8000'
-  }
+  },
+  transpileDependencies: [
+    "nimview"
+  ]
 }
