+++ src/js/nimview.js
@@ -1,161 +1,2 @@
-/** Nimview UI Library 
- * Copyright (C) 2021, by Marco Mengelkoch
- * Licensed under MIT License, see License file for more details
- * git clone https://github.com/marcomq/nimview
-**/
-
-import "whatwg-fetch"
-
-export let ui = {}
-let backend = {}
-let defaultPostTarget = ""
-let host = "" // might cause "cors" errors if defined
-ui.resolveStorage = {}
-ui.requestCounter = 0
-ui.initStarted = false
-ui.initFinished = false
-ui.initFailed = false
-
-const createRequestId = () => {
-    if (ui.requestCounter >= Number.MAX_SAFE_INTEGER-1) {
-        ui.requestCounter = 0
-    }
-    return ui.requestCounter++
-}
-
-ui.alert = function (str) {
-    if (typeof window.nimview === 'undefined') {
-      alert(str);
-    }
-    else {
-        window.nimview.alert(str);
-    }
-}
-
-ui.addRequest = (requestOrArray) => {  
-    /*global backend functions*/
-    if (typeof requestOrArray === "string") {
-        requestOrArray = [requestOrArray, ""]
-    }
-    let request = requestOrArray[0] + ""
-    let signature = requestOrArray[1]  + "" // for debugging
-    backend[request] = (async (...data) => {
-        if (typeof window.nimview == 'undefined') {
-            // http-server
-            const postData = JSON.stringify({request: request, data: data})
-            const requestOpts = { 
-                method: 'POST', // always use AJAX post for simplicity with special chars    
-                mode: 'cors',
-                cache: 'no-cache',
-                headers: {'Content-Type': 'application/json'},
-                body: postData
-            }
-            let url = request
-            if (defaultPostTarget != "") {
-                url = defaultPostTarget
-            }
-            if (ui.globalToken && (ui.globalToken.length > 0)) {
-                requestOpts.headers["global-token"] = ui.globalToken
-            }
-            if (data.length != Math.min(signature.length, signature.split(",").length) && console && console.log) {
-                console.log("Request signature might not fit: '" & request & "' signature: '" & signature & "' data: '" & JSON.stringify(data) & "'") 
-            }
-            return fetch(host + "/" + url, requestOpts).then((response) => {
-                if (response) {
-                  var globalToken = response.headers.get("global-token");
-                  if (globalToken && (globalToken.length > 0)) {
-                    ui.globalToken = globalToken;
-                    ui.lastToken = Date.now();
-                  }
-                }
-                return response.text();
-            })
-        }
-        else {
-            // webview
-            const requestId = createRequestId()
-            const postData = JSON.stringify({request: request, data: data, requestId: requestId})
-            let promise = new Promise((resolve, reject) => {
-                ui.resolveStorage[requestId] = [resolve, reject]
-                let response = window.nimview.call(postData)
-                if (typeof response !== "undefined") {
-                    // android webview
-                    resolve(response)
-                    delete ui.resolveStorage[requestId]
-                }
-            })
-            return promise
-        }
-    })
-}
-
-ui.applyResponse = (requestId, data) => {
-    if (typeof ui.resolveStorage[requestId] !== "undefined") {
-        ui.resolveStorage[requestId][0](data)
-        delete ui.resolveStorage[requestId]
-    }
-    else {
-        ui.alert("request id '" + requestId + "' not found")
-    }
-}
-ui.rejectResponse = (requestId) => {
-    if (typeof ui.resolveStorage[requestId] !== "undefined") {
-        ui.resolveStorage[requestId][1]()
-        delete ui.resolveStorage[requestId]
-    }
-}
-ui.initRequests = (async () => {
-    if (ui.initStarted == false) {
-        ui.initStarted = true
-        ui.addRequest("getGlobalToken")
-        await backend.getGlobalToken().then((resp) => {
-            let jsResp = JSON.parse(resp)
-            if (jsResp.useGlobalToken) {
-                window.setInterval(ui.getGlobalToken, 60 * 1000)
-            }
-        }).catch(function () {
-            ui.alert("getGlobalToken failed")
-            ui.initFailed = true
-        })
-        ui.addRequest("getRequests")
-        await backend.getRequests().then((resp) => {
-            let jsResp = JSON.parse(resp)
-            for (let req of jsResp) {
-                ui.addRequest(req)
-            }
-            ui.initFinished = true
-        }).catch(function () {
-            ui.alert("getRequests failed")
-            ui.initFailed = true
-        })
-    }
-})
-/**
- * This call is optional. You need to call this function if you want to 
- * immediately call a backend function and need to make sure that all functions
- * are already available. 
- * In case you need to wait for initalization, call:
- * await backend.waitInit()
- * or
- * backend.waitInit().then(backend.functionThatNeedsToWait)
- */
-backend.waitInit = () => {
-    // everything else was complicate when using webview on windows
-    return new Promise((resolve, reject) => {
-        let waitLoop = () => {
-            if (ui.initFinished) {
-                return resolve()
-            }
-            else if (ui.initFailed) {
-                return reject()
-            }
-            else {
-                setTimeout(waitLoop, 5)
-            }
-        }
-        setTimeout(waitLoop, 1)
-    })
-}
-window.ui = ui
-window.setTimeout(ui.initRequests, 1)
-export default backend
+"use strict";function e(e,t,r,o,n,i,s){try{var a=e[i](s),u=a.value}catch(e){return void r(e)}a.done?t(u):Promise.resolve(u).then(o,n)}function t(t){return function(){var r=this,o=arguments;return new Promise((function(n,i){var s=t.apply(r,o);function a(t){e(s,n,i,a,u,"next",t)}function u(t){e(s,n,i,a,u,"throw",t)}a(void 0)}))}}function r(e){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}require("core-js/modules/es.symbol.js"),require("core-js/modules/es.symbol.description.js"),require("core-js/modules/es.object.to-string.js"),require("core-js/modules/es.symbol.iterator.js"),require("core-js/modules/es.array.iterator.js"),require("core-js/modules/es.string.iterator.js"),require("core-js/modules/web.dom-collections.iterator.js"),require("core-js/modules/es.symbol.async-iterator.js"),require("core-js/modules/es.symbol.to-string-tag.js"),require("core-js/modules/es.json.to-string-tag.js"),require("core-js/modules/es.math.to-string-tag.js"),require("core-js/modules/es.object.get-prototype-of.js"),require("core-js/modules/es.function.name.js"),require("core-js/modules/es.promise.js"),require("core-js/modules/es.regexp.to-string.js"),require("core-js/modules/web.dom-collections.for-each.js"),require("core-js/modules/es.array.slice.js"),require("core-js/modules/es.number.max-safe-integer.js"),require("core-js/modules/es.number.constructor.js"),require("core-js/modules/es.regexp.exec.js"),require("core-js/modules/es.string.split.js"),require("core-js/modules/es.array.from.js"),require("core-js/modules/es.array-buffer.constructor.js"),require("core-js/modules/es.array.map.js"),require("core-js/modules/es.object.get-own-property-names.js"),require("core-js/modules/es.typed-array.uint8-array.js"),require("core-js/modules/es.typed-array.copy-within.js"),require("core-js/modules/es.typed-array.every.js"),require("core-js/modules/es.typed-array.fill.js"),require("core-js/modules/es.typed-array.filter.js"),require("core-js/modules/es.typed-array.find.js"),require("core-js/modules/es.typed-array.find-index.js"),require("core-js/modules/es.typed-array.for-each.js"),require("core-js/modules/es.typed-array.includes.js"),require("core-js/modules/es.typed-array.index-of.js"),require("core-js/modules/es.typed-array.iterator.js"),require("core-js/modules/es.typed-array.join.js"),require("core-js/modules/es.typed-array.last-index-of.js"),require("core-js/modules/es.typed-array.map.js"),require("core-js/modules/es.typed-array.reduce.js"),require("core-js/modules/es.typed-array.reduce-right.js"),require("core-js/modules/es.typed-array.reverse.js"),require("core-js/modules/es.typed-array.set.js"),require("core-js/modules/es.typed-array.slice.js"),require("core-js/modules/es.typed-array.some.js"),require("core-js/modules/es.typed-array.sort.js"),require("core-js/modules/es.typed-array.subarray.js"),require("core-js/modules/es.typed-array.to-locale-string.js"),require("core-js/modules/es.typed-array.to-string.js"),require("core-js/modules/es.array.join.js"),require("core-js/modules/web.url.js"),require("core-js/modules/es.string.replace.js"),require("core-js/modules/es.string.trim.js");var o={exports:{}};!function(e){var t=function(e){var t,o=Object.prototype,n=o.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},s=i.iterator||"@@iterator",a=i.asyncIterator||"@@asyncIterator",u=i.toStringTag||"@@toStringTag";function c(e,t,r){return Object.defineProperty(e,t,{value:r,enumerable:!0,configurable:!0,writable:!0}),e[t]}try{c({},"")}catch(e){c=function(e,t,r){return e[t]=r}}function l(e,t,r,o){var n=t&&t.prototype instanceof b?t:b,i=Object.create(n.prototype),s=new O(o||[]);return i._invoke=function(e,t,r){var o=h;return function(n,i){if(o===y)throw new Error("Generator is already running");if(o===p){if("throw"===n)throw i;return P()}for(r.method=n,r.arg=i;;){var s=r.delegate;if(s){var a=_(s,r);if(a){if(a===m)continue;return a}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(o===h)throw o=p,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);o=y;var u=f(e,t,r);if("normal"===u.type){if(o=r.done?p:d,u.arg===m)continue;return{value:u.arg,done:r.done}}"throw"===u.type&&(o=p,r.method="throw",r.arg=u.arg)}}}(e,r,s),i}function f(e,t,r){try{return{type:"normal",arg:e.call(t,r)}}catch(e){return{type:"throw",arg:e}}}e.wrap=l;var h="suspendedStart",d="suspendedYield",y="executing",p="completed",m={};function b(){}function v(){}function g(){}var w={};w[s]=function(){return this};var j=Object.getPrototypeOf,q=j&&j(j(L([])));q&&q!==o&&n.call(q,s)&&(w=q);var x=g.prototype=b.prototype=Object.create(w);function T(e){["next","throw","return"].forEach((function(t){c(e,t,(function(e){return this._invoke(t,e)}))}))}function E(e,t){function o(i,s,a,u){var c=f(e[i],e,s);if("throw"!==c.type){var l=c.arg,h=l.value;return h&&"object"===r(h)&&n.call(h,"__await")?t.resolve(h.__await).then((function(e){o("next",e,a,u)}),(function(e){o("throw",e,a,u)})):t.resolve(h).then((function(e){l.value=e,a(l)}),(function(e){return o("throw",e,a,u)}))}u(c.arg)}var i;this._invoke=function(e,r){function n(){return new t((function(t,n){o(e,r,t,n)}))}return i=i?i.then(n,n):n()}}function _(e,r){var o=e.iterator[r.method];if(o===t){if(r.delegate=null,"throw"===r.method){if(e.iterator.return&&(r.method="return",r.arg=t,_(e,r),"throw"===r.method))return m;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return m}var n=f(o,e.iterator,r.arg);if("throw"===n.type)return r.method="throw",r.arg=n.arg,r.delegate=null,m;var i=n.arg;return i?i.done?(r[e.resultName]=i.value,r.next=e.nextLoc,"return"!==r.method&&(r.method="next",r.arg=t),r.delegate=null,m):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,m)}function A(e){var t={tryLoc:e[0]};1 in e&&(t.catchLoc=e[1]),2 in e&&(t.finallyLoc=e[2],t.afterLoc=e[3]),this.tryEntries.push(t)}function S(e){var t=e.completion||{};t.type="normal",delete t.arg,e.completion=t}function O(e){this.tryEntries=[{tryLoc:"root"}],e.forEach(A,this),this.reset(!0)}function L(e){if(e){var r=e[s];if(r)return r.call(e);if("function"==typeof e.next)return e;if(!isNaN(e.length)){var o=-1,i=function r(){for(;++o<e.length;)if(n.call(e,o))return r.value=e[o],r.done=!1,r;return r.value=t,r.done=!0,r};return i.next=i}}return{next:P}}function P(){return{value:t,done:!0}}return v.prototype=x.constructor=g,g.constructor=v,v.displayName=c(g,u,"GeneratorFunction"),e.isGeneratorFunction=function(e){var t="function"==typeof e&&e.constructor;return!!t&&(t===v||"GeneratorFunction"===(t.displayName||t.name))},e.mark=function(e){return Object.setPrototypeOf?Object.setPrototypeOf(e,g):(e.__proto__=g,c(e,u,"GeneratorFunction")),e.prototype=Object.create(x),e},e.awrap=function(e){return{__await:e}},T(E.prototype),E.prototype[a]=function(){return this},e.AsyncIterator=E,e.async=function(t,r,o,n,i){void 0===i&&(i=Promise);var s=new E(l(t,r,o,n),i);return e.isGeneratorFunction(r)?s:s.next().then((function(e){return e.done?e.value:s.next()}))},T(x),c(x,u,"Generator"),x[s]=function(){return this},x.toString=function(){return"[object Generator]"},e.keys=function(e){var t=[];for(var r in e)t.push(r);return t.reverse(),function r(){for(;t.length;){var o=t.pop();if(o in e)return r.value=o,r.done=!1,r}return r.done=!0,r}},e.values=L,O.prototype={constructor:O,reset:function(e){if(this.prev=0,this.next=0,this.sent=this._sent=t,this.done=!1,this.delegate=null,this.method="next",this.arg=t,this.tryEntries.forEach(S),!e)for(var r in this)"t"===r.charAt(0)&&n.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=t)},stop:function(){this.done=!0;var e=this.tryEntries[0].completion;if("throw"===e.type)throw e.arg;return this.rval},dispatchException:function(e){if(this.done)throw e;var r=this;function o(o,n){return a.type="throw",a.arg=e,r.next=o,n&&(r.method="next",r.arg=t),!!n}for(var i=this.tryEntries.length-1;i>=0;--i){var s=this.tryEntries[i],a=s.completion;if("root"===s.tryLoc)return o("end");if(s.tryLoc<=this.prev){var u=n.call(s,"catchLoc"),c=n.call(s,"finallyLoc");if(u&&c){if(this.prev<s.catchLoc)return o(s.catchLoc,!0);if(this.prev<s.finallyLoc)return o(s.finallyLoc)}else if(u){if(this.prev<s.catchLoc)return o(s.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<s.finallyLoc)return o(s.finallyLoc)}}}},abrupt:function(e,t){for(var r=this.tryEntries.length-1;r>=0;--r){var o=this.tryEntries[r];if(o.tryLoc<=this.prev&&n.call(o,"finallyLoc")&&this.prev<o.finallyLoc){var i=o;break}}i&&("break"===e||"continue"===e)&&i.tryLoc<=t&&t<=i.finallyLoc&&(i=null);var s=i?i.completion:{};return s.type=e,s.arg=t,i?(this.method="next",this.next=i.finallyLoc,m):this.complete(s)},complete:function(e,t){if("throw"===e.type)throw e.arg;return"break"===e.type||"continue"===e.type?this.next=e.arg:"return"===e.type?(this.rval=this.arg=e.arg,this.method="return",this.next="end"):"normal"===e.type&&t&&(this.next=t),m},finish:function(e){for(var t=this.tryEntries.length-1;t>=0;--t){var r=this.tryEntries[t];if(r.finallyLoc===e)return this.complete(r.completion,r.afterLoc),S(r),m}},catch:function(e){for(var t=this.tryEntries.length-1;t>=0;--t){var r=this.tryEntries[t];if(r.tryLoc===e){var o=r.completion;if("throw"===o.type){var n=o.arg;S(r)}return n}}throw new Error("illegal catch attempt")},delegateYield:function(e,r,o){return this.delegate={iterator:L(e),resultName:r,nextLoc:o},"next"===this.method&&(this.arg=t),m}},e}(e.exports);try{regeneratorRuntime=t}catch(e){Function("r","regeneratorRuntime = r")(t)}}(o);var n=o.exports,i="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==i&&i,s="URLSearchParams"in i,a="Symbol"in i&&"iterator"in Symbol,u="FileReader"in i&&"Blob"in i&&function(){try{return new Blob,!0}catch(e){return!1}}(),c="FormData"in i,l="ArrayBuffer"in i;if(l)var f=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],h=ArrayBuffer.isView||function(e){return e&&f.indexOf(Object.prototype.toString.call(e))>-1};function d(e){if("string"!=typeof e&&(e=String(e)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(e)||""===e)throw new TypeError('Invalid character in header field name: "'+e+'"');return e.toLowerCase()}function y(e){return"string"!=typeof e&&(e=String(e)),e}function p(e){var t={next:function(){var t=e.shift();return{done:void 0===t,value:t}}};return a&&(t[Symbol.iterator]=function(){return t}),t}function m(e){this.map={},e instanceof m?e.forEach((function(e,t){this.append(t,e)}),this):Array.isArray(e)?e.forEach((function(e){this.append(e[0],e[1])}),this):e&&Object.getOwnPropertyNames(e).forEach((function(t){this.append(t,e[t])}),this)}function b(e){if(e.bodyUsed)return Promise.reject(new TypeError("Already read"));e.bodyUsed=!0}function v(e){return new Promise((function(t,r){e.onload=function(){t(e.result)},e.onerror=function(){r(e.error)}}))}function g(e){var t=new FileReader,r=v(t);return t.readAsArrayBuffer(e),r}function w(e){if(e.slice)return e.slice(0);var t=new Uint8Array(e.byteLength);return t.set(new Uint8Array(e)),t.buffer}function j(){return this.bodyUsed=!1,this._initBody=function(e){var t;this.bodyUsed=this.bodyUsed,this._bodyInit=e,e?"string"==typeof e?this._bodyText=e:u&&Blob.prototype.isPrototypeOf(e)?this._bodyBlob=e:c&&FormData.prototype.isPrototypeOf(e)?this._bodyFormData=e:s&&URLSearchParams.prototype.isPrototypeOf(e)?this._bodyText=e.toString():l&&u&&((t=e)&&DataView.prototype.isPrototypeOf(t))?(this._bodyArrayBuffer=w(e.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):l&&(ArrayBuffer.prototype.isPrototypeOf(e)||h(e))?this._bodyArrayBuffer=w(e):this._bodyText=e=Object.prototype.toString.call(e):this._bodyText="",this.headers.get("content-type")||("string"==typeof e?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):s&&URLSearchParams.prototype.isPrototypeOf(e)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},u&&(this.blob=function(){var e=b(this);if(e)return e;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){if(this._bodyArrayBuffer){var e=b(this);return e||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer))}return this.blob().then(g)}),this.text=function(){var e,t,r,o=b(this);if(o)return o;if(this._bodyBlob)return e=this._bodyBlob,t=new FileReader,r=v(t),t.readAsText(e),r;if(this._bodyArrayBuffer)return Promise.resolve(function(e){for(var t=new Uint8Array(e),r=new Array(t.length),o=0;o<t.length;o++)r[o]=String.fromCharCode(t[o]);return r.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},c&&(this.formData=function(){return this.text().then(T)}),this.json=function(){return this.text().then(JSON.parse)},this}m.prototype.append=function(e,t){e=d(e),t=y(t);var r=this.map[e];this.map[e]=r?r+", "+t:t},m.prototype.delete=function(e){delete this.map[d(e)]},m.prototype.get=function(e){return e=d(e),this.has(e)?this.map[e]:null},m.prototype.has=function(e){return this.map.hasOwnProperty(d(e))},m.prototype.set=function(e,t){this.map[d(e)]=y(t)},m.prototype.forEach=function(e,t){for(var r in this.map)this.map.hasOwnProperty(r)&&e.call(t,this.map[r],r,this)},m.prototype.keys=function(){var e=[];return this.forEach((function(t,r){e.push(r)})),p(e)},m.prototype.values=function(){var e=[];return this.forEach((function(t){e.push(t)})),p(e)},m.prototype.entries=function(){var e=[];return this.forEach((function(t,r){e.push([r,t])})),p(e)},a&&(m.prototype[Symbol.iterator]=m.prototype.entries);var q=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function x(e,t){if(!(this instanceof x))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var r,o,n=(t=t||{}).body;if(e instanceof x){if(e.bodyUsed)throw new TypeError("Already read");this.url=e.url,this.credentials=e.credentials,t.headers||(this.headers=new m(e.headers)),this.method=e.method,this.mode=e.mode,this.signal=e.signal,n||null==e._bodyInit||(n=e._bodyInit,e.bodyUsed=!0)}else this.url=String(e);if(this.credentials=t.credentials||this.credentials||"same-origin",!t.headers&&this.headers||(this.headers=new m(t.headers)),this.method=(r=t.method||this.method||"GET",o=r.toUpperCase(),q.indexOf(o)>-1?o:r),this.mode=t.mode||this.mode||null,this.signal=t.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&n)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(n),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==t.cache&&"no-cache"!==t.cache)){var i=/([?&])_=[^&]*/;if(i.test(this.url))this.url=this.url.replace(i,"$1_="+(new Date).getTime());else{this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}}function T(e){var t=new FormData;return e.trim().split("&").forEach((function(e){if(e){var r=e.split("="),o=r.shift().replace(/\+/g," "),n=r.join("=").replace(/\+/g," ");t.append(decodeURIComponent(o),decodeURIComponent(n))}})),t}function E(e,t){if(!(this instanceof E))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');t||(t={}),this.type="default",this.status=void 0===t.status?200:t.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===t.statusText?"":""+t.statusText,this.headers=new m(t.headers),this.url=t.url||"",this._initBody(e)}x.prototype.clone=function(){return new x(this,{body:this._bodyInit})},j.call(x.prototype),j.call(E.prototype),E.prototype.clone=function(){return new E(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new m(this.headers),url:this.url})},E.error=function(){var e=new E(null,{status:0,statusText:""});return e.type="error",e};var _=[301,302,303,307,308];E.redirect=function(e,t){if(-1===_.indexOf(t))throw new RangeError("Invalid status code");return new E(null,{status:t,headers:{location:e}})};var A=i.DOMException;try{new A}catch(e){(A=function(e,t){this.message=e,this.name=t;var r=Error(e);this.stack=r.stack}).prototype=Object.create(Error.prototype),A.prototype.constructor=A}function S(e,t){return new Promise((function(o,n){var s=new x(e,t);if(s.signal&&s.signal.aborted)return n(new A("Aborted","AbortError"));var a=new XMLHttpRequest;function c(){a.abort()}a.onload=function(){var e,t,r={status:a.status,statusText:a.statusText,headers:(e=a.getAllResponseHeaders()||"",t=new m,e.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(e){return 0===e.indexOf("\n")?e.substr(1,e.length):e})).forEach((function(e){var r=e.split(":"),o=r.shift().trim();if(o){var n=r.join(":").trim();t.append(o,n)}})),t)};r.url="responseURL"in a?a.responseURL:r.headers.get("X-Request-URL");var n="response"in a?a.response:a.responseText;setTimeout((function(){o(new E(n,r))}),0)},a.onerror=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},a.ontimeout=function(){setTimeout((function(){n(new TypeError("Network request failed"))}),0)},a.onabort=function(){setTimeout((function(){n(new A("Aborted","AbortError"))}),0)},a.open(s.method,function(e){try{return""===e&&i.location.href?i.location.href:e}catch(t){return e}}(s.url),!0),"include"===s.credentials?a.withCredentials=!0:"omit"===s.credentials&&(a.withCredentials=!1),"responseType"in a&&(u?a.responseType="blob":l&&s.headers.get("Content-Type")&&-1!==s.headers.get("Content-Type").indexOf("application/octet-stream")&&(a.responseType="arraybuffer")),!t||"object"!==r(t.headers)||t.headers instanceof m?s.headers.forEach((function(e,t){a.setRequestHeader(t,e)})):Object.getOwnPropertyNames(t.headers).forEach((function(e){a.setRequestHeader(e,y(t.headers[e]))})),s.signal&&(s.signal.addEventListener("abort",c),a.onreadystatechange=function(){4===a.readyState&&s.signal.removeEventListener("abort",c)}),a.send(void 0===s._bodyInit?null:s._bodyInit)}))}function O(e,t){var r="undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(!r){if(Array.isArray(e)||(r=function(e,t){if(!e)return;if("string"==typeof e)return L(e,t);var r=Object.prototype.toString.call(e).slice(8,-1);"Object"===r&&e.constructor&&(r=e.constructor.name);if("Map"===r||"Set"===r)return Array.from(e);if("Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return L(e,t)}(e))||t&&e&&"number"==typeof e.length){r&&(e=r);var o=0,n=function(){};return{s:n,n:function(){return o>=e.length?{done:!0}:{done:!1,value:e[o++]}},e:function(e){throw e},f:n}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,s=!0,a=!1;return{s:function(){r=r.call(e)},n:function(){var e=r.next();return s=e.done,e},e:function(e){a=!0,i=e},f:function(){try{s||null==r.return||r.return()}finally{if(a)throw i}}}}function L(e,t){(null==t||t>e.length)&&(t=e.length);for(var r=0,o=new Array(t);r<t;r++)o[r]=e[r];return o}S.polyfill=!0,i.fetch||(i.fetch=S,i.Headers=m,i.Request=x,i.Response=E);var P=window.ui;void 0===P&&((P={}).resolveStorage={},P.requestCounter=0,P.initStarted=!1,P.initFinished=!1,P.initFailed=!1);var B=window.backend||{},R=function(){return P.requestCounter>=Number.MAX_SAFE_INTEGER-1&&(P.requestCounter=0),P.requestCounter++};P.alert=function(e){void 0===window.nimview?alert(e):window.nimview.alert(e)},P.addRequest=function(e){"string"==typeof e&&(e=[e,""]);var r=e[0]+"",o=e[1]+"";B[r]=t(n.mark((function e(){var t,i,s,a,u,c,l,f,h,d=arguments;return n.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:for(t=d.length,i=new Array(t),s=0;s<t;s++)i[s]=d[s];if(void 0!==window.nimview){e.next=11;break}return a=JSON.stringify({request:r,data:i}),u={method:"POST",mode:"cors",cache:"no-cache",headers:{"Content-Type":"application/json"},body:a},c=r,P.globalToken&&P.globalToken.length>0&&(u.headers["global-token"]=P.globalToken),i.length!=Math.min(o.length,o.split(",").length)&&console&&console.log&&console.log(0&r&o&"' data: '"&JSON.stringify(i)&"'"),e.abrupt("return",fetch("/"+c,u).then((function(e){if(e){var t=e.headers.get("global-token");t&&t.length>0&&(P.globalToken=t,P.lastToken=Date.now())}return e.text()})));case 11:return l=R(),f=JSON.stringify({request:r,data:i,requestId:l}),h=new Promise((function(e,t){P.resolveStorage[l]=[e,t];var r=window.nimview.call(f);void 0!==r&&(e(r),delete P.resolveStorage[l])})),e.abrupt("return",h);case 15:case"end":return e.stop()}}),e)})))},P.applyResponse=function(e,t){void 0!==P.resolveStorage[e]?(P.resolveStorage[e][0](t),delete P.resolveStorage[e]):P.alert("request id '"+e+"' not found")},P.rejectResponse=function(e){void 0!==P.resolveStorage[e]&&(P.resolveStorage[e][1](),delete P.resolveStorage[e])},P.initRequests=t(n.mark((function e(){return n.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(0!=P.initStarted){e.next=8;break}return P.initStarted=!0,P.addRequest("getGlobalToken"),e.next=5,B.getGlobalToken().then((function(e){JSON.parse(e).useGlobalToken&&window.setInterval(P.getGlobalToken,6e4)})).catch((function(){P.alert("getGlobalToken failed"),P.initFailed=!0}));case 5:return P.addRequest("getRequests"),e.next=8,B.getRequests().then((function(e){var t,r=O(JSON.parse(e));try{for(r.s();!(t=r.n()).done;){var o=t.value;P.addRequest(o)}}catch(e){r.e(e)}finally{r.f()}P.initFinished=!0})).catch((function(){P.alert("getRequests failed"),P.initFailed=!0}));case 8:case"end":return e.stop()}}),e)}))),B.waitInit=function(){return new Promise((function(e,t){setTimeout((function r(){return P.initFinished?e():P.initFailed?t():void setTimeout(r,5)}),1)}))},window.ui=P,window.backend=B,window.setTimeout(P.initRequests,1),module.exports=B;
+//# sourceMappingURL=nimview.js.map
