+++ src/js/package.json
@@ -1,9 +1,10 @@
 {
   "name": "nimview",
-  "version": "0.2.2",
+  "version": "0.2.3",
   "description": "This package is part of marcomq/nimview",
   "main": "nimview.js",
   "scripts": {
+    "build": "rollup -c",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -12,6 +13,7 @@
   },
   "keywords": [
     "nim",
+    "nimview",
     "webview"
   ],
   "author": "Marco Mengelkoch",
@@ -19,6 +21,18 @@
   "homepage": "https://github.com/marcomq/nimview#readme",
   "bugs": {
     "url": "https://github.com/marcomq/nimview/issues"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.14.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.14.3",
+    "@babel/preset-env": "^7.14.4",
+    "@babel/runtime": "^7.14.0",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-node-resolve": "^13.0.0",
+    "rollup": "^2.50.4",
+    "rollup-plugin-terser": "^7.0.2"
   },
   "dependencies": {
     "whatwg-fetch": "^3.6.2"
